now has a slider for the notification range because that is easier to use on a mobile device

+++ errai-jpa/demos/errai-jpa-demo-grocery-list/src/main/java/org/jboss/errai/demo/grocery/client/shared/Store.java
@@ -21,7 +21,7 @@ public class Store {
   private double latitude;
   private double longitude;
 
-  private float radius;
+  private double radius;
 
   @OneToMany(cascade={CascadeType.PERSIST, CascadeType.MERGE})
   private List<Department> departments = new ArrayList<Department>();
@@ -58,11 +58,11 @@ public class Store {
     this.longitude = longitude;
   }
 
-  public float getRadius() {
+  public double getRadius() {
     return radius;
   }
 
-  public void setRadius(float radius) {
+  public void setRadius(double radius) {
     this.radius = radius;
   }
